refactor:기본상태입력pages slug 변경 완료  #54
