+++ omechu-app/src/app/mypage/user-info-edit/[step]/GenderStep.tsx
@@ -7,22 +7,37 @@ import ProgressBar from "@/app/components/common/ProgressBar";
 import AlertModal from "@/app/components/common/AlertModal";
 import ModalWrapper from "@/app/components/common/ModalWrapper";
 import { useOnboardingStore } from "@/lib/stores/onboarding.store";
+import { indexToSlug } from "./page";
 
 export default function GenderStep() {
   const router = useRouter();
   const [showModal, setShowModal] = useState(false);
 
-  // Zustand 스토어에서 상태와 함수 가져옴
+  // Zustand에서 상태와 초기화 함수들 가져옴
   const gender = useOnboardingStore((state) => state.gender);
   const setGender = useOnboardingStore((state) => state.setGender);
+  const resetGender = useOnboardingStore((state) => state.resetGender);
+  const resetAll = useOnboardingStore((state) => state.reset); //전체 초기화 함수
 
+  // 성별 버튼 클릭하면 토글 형식으로 선택/취소
   const handleGenderClick = (selectedGender: "남성" | "여성") => {
-    // 이미 선택한 상태에서 다시 클릭하면 취소
     if (gender === selectedGender) {
-      setGender(null);
+      setGender(null); // 이미 선택되어 있으면 취소
     } else {
-      setGender(selectedGender);
+      setGender(selectedGender); // 아니면 선택
     }
+  };
+
+  // 건너뛰기 누르면 값 초기화하고 다음 페이지로
+  const handleSkip = () => {
+    resetGender(); // 초기화 함수
+    router.push(`/mypage/user-info-edit/${indexToSlug[2]}`);
+  };
+
+  // 저장 누르면 그대로 다음 단계로 넘어감
+  const handleSave = () => {
+    if (!gender) return; // 선택 없으면 저장 안 함
+    router.push(`/mypage/user-info-edit/${indexToSlug[2]}`);
   };
 
   return (
@@ -39,7 +54,7 @@ export default function GenderStep() {
           <h1 className="text-3xl font-medium">성별은 무엇인가요?</h1>
         </section>
 
-        {/* 성별 버튼 */}
+        {/* 성별 선택 버튼 2개 */}
         <section className="my-10">
           <div className="flex gap-5">
             {["여성", "남성"].map((label) => (
@@ -54,32 +69,30 @@ export default function GenderStep() {
                   }
                 `}
               >
-                <span>{label}</span>
+                {label}
               </button>
             ))}
           </div>
         </section>
       </main>
 
-      {/* 건너뛰기 / 다음 */}
+      {/* 건너뛰기 / 저장 버튼 */}
       <footer className="flex flex-col items-end w-full pb-[env(safe-area-inset-bottom)] gap-3">
         <button
-          onClick={() => {
-            router.push("./state");
-          }}
+          onClick={handleSkip}
           className="mr-5 text-base text-[#828282] dark:text-white dark:font-semibold"
         >
           건너뛰기 {">"}
         </button>
         <button
-          onClick={() => {
-            router.push("./state");
-          }}
-          className="p-2 min-w-full h-12 rounded-t-md 
-                    text-white text-xl font-normal
-                    bg-[#1F9BDA] dark:bg-[#1774a4]
-                    hover:bg-[#1c8cc4] dark:hover:bg-[#135d83]
-                    active:bg-[#197cae] dark:active:bg-[#0e4662]"
+          onClick={handleSave}
+          disabled={!gender} // 성별 선택 안 하면 저장 비활성화
+          className={`p-2 min-w-full h-12 rounded-t-md text-white text-xl font-normal
+            ${
+              gender
+                ? "bg-[#1F9BDA] dark:bg-[#1774a4] hover:bg-[#1c8cc4] dark:hover:bg-[#135d83] active:bg-[#197cae] dark:active:bg-[#0e4662]"
+                : "bg-[#A1A1A1] dark:bg-[#555] cursor-not-allowed"
+            }`}
         >
           저장
         </button>
@@ -94,8 +107,9 @@ export default function GenderStep() {
             confirmText="그만하기"
             cancelText="돌아가기"
             onConfirm={() => {
+              resetAll();
               setShowModal(false);
-              router.push("./");
+              router.push(`/mypage/user-info-edit`); // 처음 화면으로 이동
             }}
             onClose={() => setShowModal(false)}
           />
